+++ src/pipelines/common.rs
@@ -25,6 +25,7 @@ use crate::electra::ElectraConfig;
 use crate::gpt2::Gpt2Config;
 use crate::gpt_neo::GptNeoConfig;
 use crate::longformer::LongformerConfig;
+use crate::m2m_100::M2M100Config;
 use crate::mbart::MBartConfig;
 use crate::mobilebert::MobileBertConfig;
 use crate::pegasus::PegasusConfig;
@@ -34,14 +35,15 @@ use crate::t5::T5Config;
 use crate::xlnet::XLNetConfig;
 use crate::Config;
 use rust_tokenizers::tokenizer::{
-    AlbertTokenizer, BertTokenizer, Gpt2Tokenizer, MBart50Tokenizer, MarianTokenizer,
-    MultiThreadedTokenizer, OpenAiGptTokenizer, PegasusTokenizer, ProphetNetTokenizer,
-    ReformerTokenizer, RobertaTokenizer, T5Tokenizer, Tokenizer, TruncationStrategy,
-    XLMRobertaTokenizer, XLNetTokenizer,
+    AlbertTokenizer, BertTokenizer, Gpt2Tokenizer, M2M100Tokenizer, MBart50Tokenizer,
+    MarianTokenizer, MultiThreadedTokenizer, OpenAiGptTokenizer, PegasusTokenizer,
+    ProphetNetTokenizer, ReformerTokenizer, RobertaTokenizer, T5Tokenizer, Tokenizer,
+    TruncationStrategy, XLMRobertaTokenizer, XLNetTokenizer,
 };
 use rust_tokenizers::vocab::{
-    AlbertVocab, BertVocab, Gpt2Vocab, MBart50Vocab, MarianVocab, OpenAiGptVocab, PegasusVocab,
-    ProphetNetVocab, ReformerVocab, RobertaVocab, T5Vocab, Vocab, XLMRobertaVocab, XLNetVocab,
+    AlbertVocab, BertVocab, Gpt2Vocab, M2M100Vocab, MBart50Vocab, MarianVocab, OpenAiGptVocab,
+    PegasusVocab, ProphetNetVocab, ReformerVocab, RobertaVocab, T5Vocab, Vocab, XLMRobertaVocab,
+    XLNetVocab,
 };
 use rust_tokenizers::{TokenIdsWithOffsets, TokenizedInput, TokensWithOffsets};
 use serde::{Deserialize, Serialize};
@@ -70,6 +72,7 @@ pub enum ModelType {
     Pegasus,
     GPTNeo,
     MBart,
+    M2M100,
 }
 
 /// # Abstraction that holds a model configuration, can be of any of the supported models
@@ -106,6 +109,8 @@ pub enum ConfigOption {
     GPTNeo(GptNeoConfig),
     /// MBart configuration
     MBart(MBartConfig),
+    /// M2M100 configuration
+    M2M100(M2M100Config),
 }
 
 /// # Abstraction that holds a particular tokenizer, can be of any of the supported models
@@ -136,6 +141,8 @@ pub enum TokenizerOption {
     Pegasus(PegasusTokenizer),
     /// MBart50 Tokenizer
     MBart50(MBart50Tokenizer),
+    /// M2M100 Tokenizer
+    M2M100(M2M100Tokenizer),
 }
 
 impl ConfigOption {
@@ -160,6 +167,7 @@ impl ConfigOption {
             ModelType::Longformer => ConfigOption::Longformer(LongformerConfig::from_file(path)),
             ModelType::Pegasus => ConfigOption::Pegasus(PegasusConfig::from_file(path)),
             ModelType::MBart => ConfigOption::MBart(MBartConfig::from_file(path)),
+            ModelType::M2M100 => ConfigOption::M2M100(M2M100Config::from_file(path)),
         }
     }
 
@@ -199,6 +207,9 @@ impl ConfigOption {
                 .id2label
                 .expect("No label dictionary (id2label) provided in configuration file"),
             Self::MBart(config) => config
+                .id2label
+                .expect("No label dictionary (id2label) provided in configuration file"),
+            Self::M2M100(config) => config
                 .id2label
                 .expect("No label dictionary (id2label) provided in configuration file"),
             Self::T5(_) => panic!("T5 does not use a label mapping"),
@@ -403,6 +414,26 @@ impl TokenizerOption {
                 }
                 TokenizerOption::MBart50(MBart50Tokenizer::from_file(vocab_path, lower_case)?)
             }
+            ModelType::M2M100 => {
+                if add_prefix_space.is_some() {
+                    return Err(RustBertError::InvalidConfigurationError(
+                        format!("Optional input `add_prefix_space` set to value {} but cannot be used by {:?}",
+                                add_prefix_space.unwrap(),
+                                model_type)));
+                }
+                if strip_accents.is_some() {
+                    return Err(RustBertError::InvalidConfigurationError(format!(
+                        "Optional input `strip_accents` set to value {} but cannot be used by {:?}",
+                        strip_accents.unwrap(),
+                        model_type
+                    )));
+                }
+                TokenizerOption::M2M100(M2M100Tokenizer::from_files(
+                    vocab_path,
+                    merges_path.expect("No merges specified!"),
+                    lower_case,
+                )?)
+            }
         };
         Ok(tokenizer)
     }
@@ -423,6 +454,7 @@ impl TokenizerOption {
             Self::ProphetNet(_) => ModelType::ProphetNet,
             Self::Pegasus(_) => ModelType::Pegasus,
             Self::MBart50(_) => ModelType::MBart,
+            Self::M2M100(_) => ModelType::M2M100,
         }
     }
 
@@ -520,6 +552,13 @@ impl TokenizerOption {
                 stride,
             ),
             Self::MBart50(ref tokenizer) => MultiThreadedTokenizer::encode_list(
+                tokenizer,
+                text_list,
+                max_len,
+                truncation_strategy,
+                stride,
+            ),
+            Self::M2M100(ref tokenizer) => MultiThreadedTokenizer::encode_list(
                 tokenizer,
                 text_list,
                 max_len,
@@ -629,6 +668,13 @@ impl TokenizerOption {
                 truncation_strategy,
                 stride,
             ),
+            Self::M2M100(ref tokenizer) => MultiThreadedTokenizer::encode_pair_list(
+                tokenizer,
+                text_pair_list,
+                max_len,
+                truncation_strategy,
+                stride,
+            ),
         }
     }
 
@@ -681,6 +727,9 @@ impl TokenizerOption {
             Self::MBart50(ref tokenizer) => {
                 tokenizer.encode(text_1, text_2, max_len, truncation_strategy, stride)
             }
+            Self::M2M100(ref tokenizer) => {
+                tokenizer.encode(text_1, text_2, max_len, truncation_strategy, stride)
+            }
         }
     }
 
@@ -700,6 +749,7 @@ impl TokenizerOption {
             Self::ProphetNet(ref tokenizer) => tokenizer.tokenize(text),
             Self::Pegasus(ref tokenizer) => tokenizer.tokenize(text),
             Self::MBart50(ref tokenizer) => tokenizer.tokenize(text),
+            Self::M2M100(ref tokenizer) => tokenizer.tokenize(text),
         }
     }
 
@@ -719,6 +769,7 @@ impl TokenizerOption {
             Self::ProphetNet(ref tokenizer) => tokenizer.tokenize_with_offsets(text),
             Self::Pegasus(ref tokenizer) => tokenizer.tokenize_with_offsets(text),
             Self::MBart50(ref tokenizer) => tokenizer.tokenize_with_offsets(text),
+            Self::M2M100(ref tokenizer) => tokenizer.tokenize_with_offsets(text),
         }
     }
 
@@ -744,6 +795,7 @@ impl TokenizerOption {
             }
             Self::Pegasus(ref tokenizer) => MultiThreadedTokenizer::tokenize_list(tokenizer, text),
             Self::MBart50(ref tokenizer) => MultiThreadedTokenizer::tokenize_list(tokenizer, text),
+            Self::M2M100(ref tokenizer) => MultiThreadedTokenizer::tokenize_list(tokenizer, text),
         }
     }
 
@@ -792,6 +844,9 @@ impl TokenizerOption {
                 tokenizer.decode(token_ids, skip_special_tokens, clean_up_tokenization_spaces)
             }
             Self::MBart50(ref tokenizer) => {
+                tokenizer.decode(token_ids, skip_special_tokens, clean_up_tokenization_spaces)
+            }
+            Self::M2M100(ref tokenizer) => {
                 tokenizer.decode(token_ids, skip_special_tokens, clean_up_tokenization_spaces)
             }
         }
@@ -856,6 +911,10 @@ impl TokenizerOption {
                 token_ids_with_offsets_1,
                 token_ids_with_offsets_2,
             ),
+            Self::M2M100(ref tokenizer) => tokenizer.build_input_with_special_tokens(
+                token_ids_with_offsets_1,
+                token_ids_with_offsets_2,
+            ),
         };
         TokenizedInput {
             token_ids: token_ids_with_special_tokens.token_ids,
@@ -889,6 +948,7 @@ impl TokenizerOption {
             Self::ProphetNet(ref tokenizer) => tokenizer.convert_tokens_to_ids(tokens),
             Self::Pegasus(ref tokenizer) => tokenizer.convert_tokens_to_ids(tokens),
             Self::MBart50(ref tokenizer) => tokenizer.convert_tokens_to_ids(tokens),
+            Self::M2M100(ref tokenizer) => tokenizer.convert_tokens_to_ids(tokens),
         }
     }
 
@@ -946,6 +1006,10 @@ impl TokenizerOption {
             Self::MBart50(ref tokenizer) => *MultiThreadedTokenizer::vocab(tokenizer)
                 .special_values
                 .get(MBart50Vocab::unknown_value())
+                .expect("UNK token not found in vocabulary"),
+            Self::M2M100(ref tokenizer) => *MultiThreadedTokenizer::vocab(tokenizer)
+                .special_values
+                .get(M2M100Vocab::unknown_value())
                 .expect("UNK token not found in vocabulary"),
         }
     }
@@ -1013,6 +1077,12 @@ impl TokenizerOption {
                     .get(MBart50Vocab::pad_value())
                     .expect("PAD token not found in vocabulary"),
             ),
+            Self::M2M100(ref tokenizer) => Some(
+                *MultiThreadedTokenizer::vocab(tokenizer)
+                    .special_values
+                    .get(M2M100Vocab::pad_value())
+                    .expect("PAD token not found in vocabulary"),
+            ),
             Self::Reformer(_) => None,
             Self::GPT2(_) => None,
             Self::OpenAiGpt(_) => None,
@@ -1062,6 +1132,12 @@ impl TokenizerOption {
                 *MultiThreadedTokenizer::vocab(tokenizer)
                     .special_values
                     .get(MBart50Vocab::sep_value())
+                    .expect("SEP token not found in vocabulary"),
+            ),
+            Self::M2M100(ref tokenizer) => Some(
+                *MultiThreadedTokenizer::vocab(tokenizer)
+                    .special_values
+                    .get(M2M100Vocab::sep_value())
                     .expect("SEP token not found in vocabulary"),
             ),
             Self::Marian(_) => None,